agents/agent3/agent_worker/requirements-base.txt
psycopg2-binary>=2.9.0
pymongo>=4.0.0
minio>=7.0.0
python-dotenv>=1.0.0
playwright>=1.40.0

